+++ pizzaordersystem/pom.xml
@@ -63,6 +63,18 @@
 			<version>1.18.20</version>
 			<scope>provided</scope>
 		</dependency>
+				<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>slf4j-simple</artifactId>
+			<version>2.0.0-alpha0</version>
+			<scope>runtime</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>slf4j-api</artifactId>
+			<version>2.0.0-alpha0</version>
+		</dependency>
+
 
 	</dependencies>
